Move update attributes to shared
