+++ packages/block-library/src/navigation-link/shared/update-attributes.js
